Port the sensors/tracing package on Windows.
Add platform specific and platform common code in sensors/tracing package
